requirements.txt
asgiref
certifi
charset-normalizer
Django
djangorestframework
idna
phonenumbers
pillow
psycopg2
python-dotenv
requests
sqlparse
tzdata
urllib3
gunicorn